// src/types/appointments_api/responses/IRescheduleAppointmentResponse.ts
export default interface IRescheduleAppointmentResponse {
    patientId: string;
    patientFullName: string;
    patientPhoneNumber: string;
    patientDateOfBirth: string;

    doctorId: string;
    doctorFullName: string;

    specializationId: string;
    doctorSpecializationName: string;

    serviceId: string;
    serviceName: string;
    duration: number;

    officeId: string;
    officeAddress: string;

    date: string;
    time: string;
}